+++ packages/dd-trace/src/remote_config/capabilities.js
