+++ packages/dd-trace/src/remote_config/index.js
@@ -1,10 +1,10 @@
 'use strict'
 
-const Activation = require('../activation')
+const Activation = require('../appsec/activation')
 
 const RemoteConfigManager = require('./manager')
 const RemoteConfigCapabilities = require('./capabilities')
-const apiSecuritySampler = require('../api_security_sampler')
+const apiSecuritySampler = require('../appsec/api_security_sampler')
 
 let rc
 
@@ -63,7 +63,7 @@ function enableOrDisableAppsec (action, rcConfig, config, appsec) {
 function enableWafUpdate (appsecConfig) {
   if (rc && appsecConfig && !appsecConfig.rules) {
     // dirty require to make startup faster for serverless
-    const RuleManager = require('../rule_manager')
+    const RuleManager = require('../appsec/rule_manager')
 
     rc.updateCapabilities(RemoteConfigCapabilities.ASM_IP_BLOCKING, true)
     rc.updateCapabilities(RemoteConfigCapabilities.ASM_USER_BLOCKING, true)
@@ -86,7 +86,7 @@ function enableWafUpdate (appsecConfig) {
 
 function disableWafUpdate () {
   if (rc) {
-    const RuleManager = require('../rule_manager')
+    const RuleManager = require('../appsec/rule_manager')
 
     rc.updateCapabilities(RemoteConfigCapabilities.ASM_IP_BLOCKING, false)
     rc.updateCapabilities(RemoteConfigCapabilities.ASM_USER_BLOCKING, false)
